RAG/backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure OpenAI
openai==1.3.5
tiktoken==0.5.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Data Processing
pypdf==3.17.1
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
